Upates a comment in the emacs config file.
The setenv command only works for eshell. So setting $VISUAL in
the config file does not set it when the shell is bash. This puts a
damper on my usage of emacsclient as $EDITOR/$VISUAL. It will just
have to be set by the user in their bashrc file (or however they set
environment variables).