Makes start-shell an interactive function.
Now you can call M-x start-shell which will prompt you for a new
buffer name to create a shell in. This is the same function that
the two shells created on startup use.

I found myself occasionally needing a third shell when 1 & 2 were
already in use. Rather than using M-x eval-expression and calling the
shell lisp function, this is easier.