Added clang-format call as well as a few cleanup changes.
I added the clang-format el file to be able to use
clang-format-region for c++ code. I also moved the packages
to the emacs.d directory which should make installing this
config simpler.

For the clang-format call, the clang-format tool needs to be
installed separately. I included some instructions in the emacs
file for ubuntu.

I also turned on showing the col number for a line.